+++ packages/uni-app/App.vue
@@ -28,9 +28,6 @@ export default {
 	},
 	onShow: function () {
 		console.log('App Show')
-		// 应用从后台回到前台时同步路由状态
-		// 注意：页面切换时的状态同步应在各页面的 onShow 中调用 syncRoute()
-		router.syncRoute()
 	},
 	onHide: function () {
 		console.log('App Hide')