+++ packages/uni-app/uni_modules/vite-plugin/js_sdk/common/script/index.mjs
@@ -1,0 +1,1 @@
+function o(c,e="callback",r=""){return c?`function(${r}) { try { ${c} } catch(e) { console.error('[${e}] error:', e); } }`:`function(${r}) {}`}export{o as makeCallback};